// frontend-recode/src/types/impl/api/static/operator.ts
export type Operator = {
// ...
    phases: {
        characterPrefabKey: string;
        rangeId: string;
        maxLevel: number;
        attributesKeyFrames: {
            level: number;
            data: {
                maxHp: number;
                atk: number;
                def: number;
                magicResistance: number;
                cost: number;
                blockCnt: number;
                moveSpeed: number;
                attackSpeed: number;
                baseAttackTime: number;
                respawnTime: number;
                hpRecoveryPerSec: number;
                spRecoveryPerSec: number;
                maxDeployCount: number;
                maxDeckStackCnt: number;
                tauntLevel: number;
                massLevel: number;
                baseForceLevel: number;
                stunImmune: boolean;
                silenceImmune: boolean;
                sleepImmune: boolean;
                frozenImmune: boolean;
                levitateImmune: boolean;
                disarmedCombatImmune: boolean;
            };
        }[];
        evolveCost:
            | {
                  id: string;
                  count: number;
                  type: ItemType;
              }[]
            | null;
    }[];
// ...
};

export enum OperatorPosition {
    RANGED = "RANGED",
    MELEE = "MELEE",
}

export enum OperatorPhase {
    ELITE_0 = "PHASE_0",
    ELITE_1 = "PHASE_1",
    ELITE_2 = "PHASE_2",
}

export enum OperatorRarity {
    sixStar = "TIER_6",
    fiveStar = "TIER_5",
    fourStar = "TIER_4",
    threeStar = "TIER_3",
    twoStar = "TIER_2",
    oneStar = "TIER_1",
}